support upsert record _id for Mongo 2.4
NTFY-179: This fixes a regression from a previous change where the _id
property was only included in the query but not in the record during an
upserted update.  In Mongo 2.6 this works fine and the query _id is
used as the record _id but in Mongo 2.4, the _id gets newly created as
an ObjectId().